requirements.txt
numpy==1.24.3
pandas==1.5.3
tensorflow==2.11.0
scikit-learn==1.2.2
python-dotenv==1.0.0
requests==2.28.1
